requirements.txt
langchain
langgraph
langsmith
qdrant-client
streamlit
openai
python-dotenv
PyPDF2
pytest
langchain_groq